Move the navbar contextual components down as nested components.
