+++ addon/components/bs-navbar/nav.js
@@ -1,5 +1,5 @@
 import BsNavComponent from 'ember-bootstrap/components/bs-nav';
-import layout from '../templates/components/bs-navbar-nav';
+import layout from '../../templates/components/bs-navbar/nav';
 
 /**
  * Component for the `.nav` element within a [Components.Navbar](Components.Navbar.html)
