+++ addon/templates/components/bs-navbar.hbs
@@ -1,9 +1,9 @@
 <div class={{if fluid "container-fluid" "container"}}>
     {{yield
         (hash
-            toggle=(component 'bs-navbar-toggle' onClick=(action 'toggleNavbar') collapsed=collapsed)
-            content=(component 'bs-navbar-content' collapsed=collapsed)
-            nav=(component 'bs-navbar-nav')
+            toggle=(component 'bs-navbar/toggle' onClick=(action 'toggleNavbar') collapsed=collapsed)
+            content=(component 'bs-navbar/content' collapsed=collapsed)
+            nav=(component 'bs-navbar/nav')
             nav-item=(component 'bs-nav-item')
         )
     }}
