+++ addon/templates/components/bs-navbar/content.hbs
