+++ addon/templates/components/bs-navbar/nav.hbs
