+++ addon/templates/components/bs-navbar/toggle.hbs
