--- app/components/bs-navbar-content.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-bootstrap/components/bs-navbar-content';
