--- app/components/bs-navbar-nav.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-bootstrap/components/bs-navbar-nav';
