--- app/components/bs-navbar-toggle.js
@@ -1,1 +1,0 @@
-export { default } from 'ember-bootstrap/components/bs-navbar-toggle';
