+++ app/components/bs-navbar/nav.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-bootstrap/components/bs-navbar/nav';
