+++ app/components/bs-navbar/toggle.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-bootstrap/components/bs-navbar/toggle';
