+++ tests/integration/components/bs-navbar/content-test.js
@@ -1,20 +1,20 @@
 import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 
-moduleForComponent('bs-navbar-content', 'Integration | Component | bs-navbar-content', {
+moduleForComponent('bs-navbar/content', 'Integration | Component | bs-navbar/content', {
   integration: true
 });
 
 test('it renders', function(assert) {
-  this.render(hbs`{{bs-navbar-content}}`);
+  this.render(hbs`{{bs-navbar/content}}`);
 
   assert.equal(this.$().text().trim(), '');
 
   // Template block usage:
   this.render(hbs`
-    {{#bs-navbar-content}}
+    {{#bs-navbar/content}}
       template block text
-    {{/bs-navbar-content}}
+    {{/bs-navbar/content}}
   `);
 
   assert.equal(this.$().text().trim(), 'template block text');
@@ -22,9 +22,9 @@ test('it renders', function(assert) {
 
 test('it has correct markup', function(assert) {
   this.render(hbs`
-    {{#bs-navbar-content}}
+    {{#bs-navbar/content}}
       Content
-    {{/bs-navbar-content}}
+    {{/bs-navbar/content}}
   `);
 
   assert.equal(this.$().text().trim(), 'Content');
